// ~/Projects/com.gitlab.yunieralvarez.expense-tracker/expense-tracker/ts/packages/expense-tracker-api/src/app/config/enums/config-key.enum.ts
export enum ConfigKey {
	AppEnv = 'APP_ENV',
	AppPlaidClientName = 'APP_PLAID_CLIENT_NAME',
	AppPlaidTransactionSyncCount = 'APP_PLAID_TRANSACTION_SYNC_COUNT',
	AppPlaidTransactionSyncOnFailureDelay = 'APP_PLAID_TRANSACTION_SYNC_ON_FAILURE_DELAY',
	AppPort = 'APP_PORT',
	JwtAccessTokenTtl = 'JWT_ACCESS_TOKEN_TTL',
	JwtRefreshTokenTtl = 'JWT_REFRESH_TOKEN_TTL',
	JwtSecret = 'JWT_SECRET',
	JwtTokenAudience = 'JWT_TOKEN_AUDIENCE',
	JwtTokenIssuer = 'JWT_TOKEN_ISSUER',
	PlaidClientId = 'PLAID_CLIENT_ID',
	PlaidCountryCodes = 'PLAID_COUNTRY_CODES',
	PlaidEnv = 'PLAID_ENV',
	PlaidProducts = 'PLAID_PRODUCTS',
	PlaidRedirectUri = 'PLAID_REDIRECT_URI',
	PlaidSecret = 'PLAID_SECRET',
	PlaidVersion = 'PLAID_VERSION',
	PostgresDatabase = 'POSTGRES_DATABASE',
	PostgresHost = 'POSTGRES_HOST',
	PostgresPassword = 'POSTGRES_PASSWORD',
	PostgresPort = 'POSTGRES_PORT',
	PostgresUsername = 'POSTGRES_USERNAME',
}